elearning/requirements.txt
asgiref==3.8.1
bleach==6.2.0
dj-database-url==2.2.0
Django==5.1.4
django-environ==0.11.2
django-filter==23.1
django-ratelimit==4.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
packaging==24.1
Pillow==9.5.0
psycopg2==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2024.1
setuptools==75.6.0
six==1.16.0
sqlparse==0.5.3
tzdata==2024.1
vulture==2.14
webencodings==0.5.1
whitenoise==6.6.0
